requirements.txt
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
build==1.0.3
CacheControl==0.13.1
category-encoders==2.6.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
crashtest==0.4.1
cycler==0.12.1
Cython==3.0.9
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.8
dulwich==0.21.7
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
Flask==3.0.2
fonttools==4.50.0
fqdn==1.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imbalanced-learn==0.12.0
importlib-metadata==7.0.1
installer==0.7.0
ipykernel==6.29.4
ipython==8.22.2
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kaleido==0.2.1
keyring==24.3.0
kiwisolver==1.4.5
lightgbm==4.3.0
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==10.1.0
MouseInfo==0.1.3
msgpack==1.0.7
mss==9.0.1
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
notebook_shim==0.2.4
Nuitka==1.8.5
numba==0.59.1
numpy==1.26.4
ordered-set==4.1.0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
plotly==5.20.0
plotly-resampler==0.10.0
pmdarima==2.0.4
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycaret==3.3.0
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyod==1.1.3
pyparsing==3.1.2
pyperclip==1.8.2
pyproject_hooks==1.0.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytweening==1.0.7
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.6.0
referencing==0.34.0
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
schemdraw==0.15
scikit-base==0.7.5
scikit-learn==1.4.1.post1
scikit-plot==0.3.7
scipy==1.11.4
Send2Trash==1.8.2
shellingham==1.5.4
shiboken6==6.6.0
six==1.16.0
sktime==0.28.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
tbats==1.1.3
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.4.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
trove-classifiers==2023.11.29
tsdownsample==0.1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.1.0
virtualenv==20.25.0
wakepy==0.7.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
xxhash==3.4.1
yellowbrick==1.5
zipp==3.17.0
zstandard==0.21.0
